Create placeholder game page
