+++ src/App.tsx
@@ -4,6 +4,7 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Main from './pages/Main';
 import About from './pages/About';
 import Lobby from './pages/Lobby';
+import Game from './pages/Game';
 
 import { CurrentUser } from './types';
 import { WebSocketConnection } from './modules/ws';
@@ -19,6 +20,7 @@ function App(): JSX.Element {
         <Route exact path="/" component={() => <Main setWs={setWs} setCurrentUser={setCurrentUser} ws={ws} />} />
         <Route exact path="/aboutUs" component={About} />
         <Route exact path="/lobby" component={() => <Lobby ws={ws} currentUser={currentUser} />} />
+        <Route exact path="/game" component={() => <Game />} />
       </Switch>
     </Router>
   );
