+++ src/pages/Game.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const GamePage = (): JSX.Element => {
+  console.log('this is the game page');
+  return (
+    <div>
+      <p>This is the game page.</p>
+    </div>
+  );
+};
+
+export default GamePage;
